[batch][dag6] Include first 500 character of HTTP body on errors (#4792)
* [batch][dag6] Include first 500 character of HTTP body on errors

* move server to subfolder

* pull globals out to a file (#4794)
