+++ batch/batch/server/__init__.py
@@ -1,0 +1,5 @@
+from .server import serve
+
+__all__ = [
+    'serve'
+]
